Поменял int на long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {10, 115, 4, 1101};
+        long[] sales = {10, 115, 4, 1101};
         SalesManager salesManager = new SalesManager(sales);
         System.out.println(salesManager.max());
     }
